requirements.txt
playwright>=1.40.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
google-generativeai>=0.3.0
pypdf2>=3.0.0
python-docx>=1.0.0
pdfplumber>=0.10.0
pyyaml>=6.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
